Refactoring and tidying up _helpers/util
