+++ pycompilation/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.3.2'
+__version__ = '0.3.3'
 
 from .compilation import (
     compile_sources, link_py_so, src2obj,
